requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
fonttools==4.51.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
language-tool-python==2.8
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.10
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
